Add a tool for determining the minimum overlap for a set of labels.
This commit adds cp6_minclique.py, which takes an image table and
a value N, and determines which set of N labels have the least overlap
in the table. It also optionally takes a list of N labels as an argument
and reports which image IDs match those labels.

For example, for N==4 and round_2_public:

python2.7 cp6_minclique.py label_table.txt image_table.txt 4 | sort -n | head -1

...reports (on a single line, split for clarity)

0.0348947839688
  ['car(3):304', 'dog(5):179', 'food(8):262', 'lake(10):205']
  ['3:5 = 4/483 (0.008282)',
   '3:8 = 2/566 (0.003534)',
   '3:10 = 2/509 (0.003929)',
   '5:8 = 5/441 (0.011338)',
   '5:10 = 3/384 (0.007812)',
   '8:10 = 0/467 (0.000000)']

indicating the set car/dog/food/lake has the minimum label overlap;
that the labels occur 304,179,262,205 times respectively, and that
the confusion between them is 4/(304+179) = 4/483 for car/dog, etc.
The overlap metric is the sum of these factors.

For the record, the *maximum* overlap is:

1.62738107085
  ['female(6):1316', 'male(11):1485', 'people(13):2310', 'portrait(15):771']
  ['6:11 = 595/2801 (0.212424)',
   '6:13 = 1316/3626 (0.362934)',
   '6:15 = 482/2087 (0.230954)',
   '11:13 = 1485/3795 (0.391304)',
   '11:15 = 405/2256 (0.179521)',
   '13:15 = 771/3081 (0.250243)']

If you add an optional fourth argument of comma-separated label IDs
(e.g. '3,5,8,10'), instead of reporting clique metrics, it will output
lines in the format

image_id label_1 [label_2...]

...where label_1, etc are from the set specified in the fourth argument.
This allows you to identify which images in the image_table make up the
specified clique.